requirements.txt
langchain
langchain-openai
langgraph
tavily-python
streamlit
chromadb
python-dotenv
pydantic
langsmith
tiktoken
beautifulsoup4
lxml
requests
